# Add verbose option to pytorch hub models