@haimat 2021-04-25
@NanoCode012 Well, yes, partly. It would be great though to have an option "silent" or something that would completely remove *all* output ...