@NanoCode012 2021-04-25
I see! I understood you incorrectly. I thought you only wanted the network outputs and autoshape gone.