@glenn-jocher 2021-04-25
@NanoCode012 thanks for the PR! @haimat the implementation does appear to remove all output. The output shown is due to the .print() call, which would be called by a user by definition to print results to screen.