@haimat 2021-04-25
Ahh I see, thanks. When can this be tested via Hub?