@glenn-jocher 2021-04-25
@haimat after it's merged it would appear in hub (use force_reload=True to update your cache). Before merge you can test like this:
```
git clone https://github.com/NanoCode012/yolov5 -b hub-verbose
cd yolov5
python hubconf.py
```

One item that still prints in the .save() statement.