@glenn-jocher 2021-04-25
@NanoCode012 I tried the fix with PyTorch Hub, but unfortunately logging seems to not respond the verbose argument. It seems to behave differently when running hubconf.py directly than when loading up a hub model. I'm not sure what's causing this.