@NanoCode012 2021-04-26
Hi @glenn-jocher , I'll have to do some guessing in a few days to see. It's harder since pytorch hub models can't be tested locally. I'll check out if this is a limitation of logging with the hub models or my code and (hopefully) find a fix.