@glenn-jocher 2021-04-26
@NanoCode012 I know, they can be hard to debug. Thanks for the effort though!