@haimat 2021-04-29
I just tried it - first I did a `force_reload=True` to reload the model from the hub, then I added a `verbose=False`. However, the output was the same, incl. model layout summary. Is this update even online in the hub? Moreover, I am using my own model checkpoint like this:

    torch.hub.load("ultralytics/yolov5", "custom", path_or_model=model_path, verbose=False)

Should this new verbose option work in this case as well?