@glenn-jocher 2021-05-11
@haimat yes I've updated PyTorch Hub model loading to avoid having to build each model from scratch if no changes are asked for, so this simplifies the loading process and reduces the printed lines to screen now under most conditions. 

If you request a modification to the model it is still created first and then weights transferred, i.e. `classes=10`.

**This is not a solution to the `verbose=False` not working problem**, but it helps of course.